+++ example_depth_cache.py
@@ -39,7 +39,7 @@ import os
 import time
 
 logging.getLogger("unicorn_binance_local_depth_cache")
-logging.basicConfig(level=logging.INFO,
+logging.basicConfig(level=logging.DEBUG,
                     filename=os.path.basename(__file__) + '.log',
                     format="{asctime} [{levelname:8}] {process} {thread} {module}: {message}",
                     style="{")
@@ -47,16 +47,17 @@ logging.basicConfig(level=logging.INFO,
 market = 'BTCUSDT'
 
 # ubldc = BinanceLocalDepthCacheManager(exchange="binance.com")
-ubldc = BinanceLocalDepthCacheManager(exchange="binance.com-testnet")
+# ubldc = BinanceLocalDepthCacheManager(exchange="binance.com-testnet")
+ubldc = BinanceLocalDepthCacheManager(exchange="binance.com-futures")
 
 ubldc.create_depth_cache(markets=market)
 
 while True:
+    time.sleep(1)
     print(f"is_synchronized: {ubldc.is_depth_cache_synchronized(market)}")
     try:
         print(f"Top 10 asks: {ubldc.get_asks(market=market)[:10]}")
         print(f"Top 10 bids: {ubldc.get_bids(market=market)[:10]}")
     except DepthCacheOutOfSync as error_msg:
         print(f"ERROR: {error_msg}")
-    time.sleep(1)